remove unused blocks; fix locations; adjust timing constraints pt3
Decongest the critical path by culling some blocks that are
no longer used. Locations of all the CSR and IRQs are fixed
in this commit as well, so that we can freely remove blocks
without shifting all of the locations of CSRs thus avoiding breakage
in cross-genertaion firmware updates.

Some timing constraints are adjusted again, still exploring which
constraints are incorrectly defined or under-defined thus leading
to some bad route solutions on occassion.