fix proc_macro brokenness in rust 1.70
rust 1.70 has introduced some sort of bug where proc_macros don't
compile right. It manifests as a `cc` error where it can't
find link symbols, and the dump shows that it is trying to link
both x86 and riscv files together using an x86 linker. So,
somehow, the proc_macro subsystem is getting confused on a multi-target
system.

Turns out I think we don't need this proc_macro. it's...some fanciness
that was introduced years ago to create a random test ID that
I don't think we've ever relied upon. So I have excised this macro
and changed the calling API so we don't rely up on it.

Hopefully this fixes the CI build.